more project restructuring
