+++ ab-tester/src/db.rs
@@ -2,5 +2,3 @@ pub mod appdb;
 pub mod dbcursor;
 pub mod model;
 pub mod schema;
-pub mod stacks_burnstate_db;
-pub mod stacks_headers_db;
