+++ ab-tester/src/db/model.rs
@@ -1,7 +1,2 @@
 /// Types for this application.
 pub mod app_db;
-/// Types used for reading from a chainstate MARF.
-pub mod chainstate_db;
-/// Types used for reading from a chainstate Clarity MARF'd database.
-pub mod clarity_db;
-pub mod sortition_db;
