+++ ab-tester/src/db/schema.rs
@@ -1,7 +1,2 @@
 /// Tables for this application.
 pub mod appdb;
-/// Tables from the chainstate Sqlite database.
-pub mod chainstate_marf;
-/// Tables from the Clarity Sqlite database.
-pub mod clarity_marf;
-pub mod sortition;
