+++ ab-tester/src/environments/stacks_node/db/schema.rs
@@ -1,0 +1,3 @@
+pub mod chainstate;
+pub mod clarity;
+pub mod sortition;
