+++ ab-tester/src/environments/stacks_node/db/schema/clarity.rs
